Refactor code that registers and dumps bytecode indexes which sped up lines fall to
* move to separate files
* replace with namespaces (class with all static members is overhead)
* replace free() with JVMTI Deallocate call
